+++ public/qna.html
@@ -8,10 +8,11 @@
     <link rel="stylesheet" href="styles/reset.css">
     <link rel="stylesheet" href="./styles/qna.css">
     <title>Ajax Test</title>
-    <script src="qna.js" defer></script>
-</head>
 
-<body>
+    <script type="module" src="init.js"></script>
+  </head>
+
+  <body>
     <header id="main-header">
       <h1>STACK OOOOVERFLOW &#x1f64b;</h1>
       <div id="login-btn" class="login-btn">로그인</div>
