+++ public/urlConstant.js
@@ -1,0 +1,6 @@
+export const URL = {
+  INIT: "http://localhost:3000/api/questions",
+  LOGIN: "http://localhost:3000/api/login",
+  TOKEN_VALIDATION: "http://localhost:3000/api/token-validation",
+  ADD_REPLY: "/api/questions/:questionid/answers"
+};
