+++ public/util.js
@@ -1,0 +1,3 @@
+//util
+export const $ = document.querySelector.bind(document);
+export const $$ = document.querySelectorAll.bind(document);
